refactor: ship InfoPanel.StopWatch as bundled plugin with in-card hotkeys
- Split Stopwatch into InfoPanel.StopWatch (v1.0.2 by fweepa); remove from Extras.
- Build/copy/publish and whitelist bundled folder next to InfoPanel.Extras.
- Move global hotkey UI into PluginProperties for StopWatch only; drop USB Panels block.
- PluginsViewModel: ShowStopwatchHotkeys; fix PluginsPage Components using for PluginDetailControl.

Made-with: Cursor
