+++ InfoPanel.StopWatch/PluginInfo.ini
@@ -1,0 +1,5 @@
+[PluginInfo]
+Name=InfoPanel.StopWatch
+Description=Stopwatch with start, stop, reset and elapsed sensors for panels.
+Author=fweepa
+Version=1.0.2
