+++ InfoPanel.sln
@@ -1,4 +1,4 @@
-﻿
+
 Microsoft Visual Studio Solution File, Format Version 12.00
 # Visual Studio Version 17
 VisualStudioVersion = 17.6.33815.320
@@ -12,6 +12,8 @@ EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "InfoPanel.Plugins.Loader", "InfoPanel.Plugins.Loader\InfoPanel.Plugins.Loader.csproj", "{299B147B-6B0C-430F-9B91-AC1FB80AAF95}"
 EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "InfoPanel.Extras", "InfoPanel.Extras\InfoPanel.Extras.csproj", "{53662165-B047-4425-92F9-991221173CF4}"
+EndProject
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "InfoPanel.StopWatch", "InfoPanel.StopWatch\InfoPanel.StopWatch.csproj", "{D4E5F6A7-B8C9-4012-DEF0-123456789ABC}"
 EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "InfoPanel.Plugins.Ipc", "InfoPanel.Plugins.Ipc\InfoPanel.Plugins.Ipc.csproj", "{A1B2C3D4-E5F6-7890-ABCD-EF1234567890}"
 EndProject
@@ -47,6 +49,10 @@ Global
 		{53662165-B047-4425-92F9-991221173CF4}.Debug|x64.Build.0 = Debug|x64
 		{53662165-B047-4425-92F9-991221173CF4}.Release|x64.ActiveCfg = Release|x64
 		{53662165-B047-4425-92F9-991221173CF4}.Release|x64.Build.0 = Release|x64
+		{D4E5F6A7-B8C9-4012-DEF0-123456789ABC}.Debug|x64.ActiveCfg = Debug|x64
+		{D4E5F6A7-B8C9-4012-DEF0-123456789ABC}.Debug|x64.Build.0 = Debug|x64
+		{D4E5F6A7-B8C9-4012-DEF0-123456789ABC}.Release|x64.ActiveCfg = Release|x64
+		{D4E5F6A7-B8C9-4012-DEF0-123456789ABC}.Release|x64.Build.0 = Release|x64
 		{A1B2C3D4-E5F6-7890-ABCD-EF1234567890}.Debug|x64.ActiveCfg = Debug|x64
 		{A1B2C3D4-E5F6-7890-ABCD-EF1234567890}.Debug|x64.Build.0 = Debug|x64
 		{A1B2C3D4-E5F6-7890-ABCD-EF1234567890}.Release|x64.ActiveCfg = Release|x64
