+++ InfoPanel/InfoPanel.csproj
@@ -1,4 +1,4 @@
-﻿<Project Sdk="Microsoft.NET.Sdk">
+<Project Sdk="Microsoft.NET.Sdk">
 
 	<PropertyGroup>
 		<OutputType>WinExe</OutputType>
@@ -23,42 +23,55 @@
 		</PropertyGroup>
 	</Target>
 
-	<!-- Build InfoPanel.Extras and get its output path -->
+	<!-- Build bundled plugin assemblies and resolve output directories -->
 	<Target Name="BuildExtras" BeforeTargets="Build">
 		<MSBuild Projects="..\InfoPanel.Extras\InfoPanel.Extras.csproj" Targets="Build" Properties="Configuration=$(Configuration);Platform=$(Platform);TargetFramework=net8.0-windows;RuntimeIdentifier=win-x64;SelfContained=false" />
-		<!-- Get the output path after building -->
+		<MSBuild Projects="..\InfoPanel.StopWatch\InfoPanel.StopWatch.csproj" Targets="Build" Properties="Configuration=$(Configuration);Platform=$(Platform);TargetFramework=net8.0-windows;RuntimeIdentifier=win-x64;SelfContained=false" />
 		<MSBuild Projects="..\InfoPanel.Extras\InfoPanel.Extras.csproj" Targets="GetTargetPath" Properties="Configuration=$(Configuration);Platform=$(Platform);TargetFramework=net8.0-windows;RuntimeIdentifier=win-x64">
 			<Output TaskParameter="TargetOutputs" PropertyName="_ExtrasAssemblyPath" />
 		</MSBuild>
+		<MSBuild Projects="..\InfoPanel.StopWatch\InfoPanel.StopWatch.csproj" Targets="GetTargetPath" Properties="Configuration=$(Configuration);Platform=$(Platform);TargetFramework=net8.0-windows;RuntimeIdentifier=win-x64">
+			<Output TaskParameter="TargetOutputs" PropertyName="_StopWatchAssemblyPath" />
+		</MSBuild>
 		<PropertyGroup>
 			<ExtrasOutputDir>$([System.IO.Path]::GetDirectoryName('$(_ExtrasAssemblyPath)'))</ExtrasOutputDir>
+			<StopWatchOutputDir>$([System.IO.Path]::GetDirectoryName('$(_StopWatchAssemblyPath)'))</StopWatchOutputDir>
 		</PropertyGroup>
 		<Message Text="InfoPanel.Extras output directory: $(ExtrasOutputDir)" Importance="High" />
+		<Message Text="InfoPanel.StopWatch output directory: $(StopWatchOutputDir)" Importance="High" />
 	</Target>
 
-	<!-- Copy InfoPanel.Extras to build output plugin dir -->
+	<!-- Copy bundled plugins to build output plugin dirs -->
 	<Target Name="CopyExtrasToBuildOutput" AfterTargets="BuildExtras">
 		<MakeDir Directories="$(OutputPath)plugins" />
 		<ItemGroup>
 			<ExtrasFiles Include="$(ExtrasOutputDir)\**\*.*" />
+			<StopWatchFiles Include="$(StopWatchOutputDir)\**\*.*" />
 		</ItemGroup>
 		<Copy SourceFiles="@(ExtrasFiles)" DestinationFolder="$(OutputPath)\plugins\InfoPanel.Extras" SkipUnchangedFiles="true" />
+		<Copy SourceFiles="@(StopWatchFiles)" DestinationFolder="$(OutputPath)\plugins\InfoPanel.StopWatch" SkipUnchangedFiles="true" />
 	</Target>
 
-	<!-- Copy InfoPanel.Extras to publish dir -->
+	<!-- Copy bundled plugins to publish dir -->
 	<Target Name="CopyExtrasToPublishOutput" AfterTargets="Publish">
-		<!-- Get the output path for copying -->
 		<MSBuild Projects="..\InfoPanel.Extras\InfoPanel.Extras.csproj" Targets="GetTargetPath" Properties="Configuration=$(Configuration);Platform=$(Platform);TargetFramework=net8.0-windows;RuntimeIdentifier=win-x64">
 			<Output TaskParameter="TargetOutputs" PropertyName="_ExtrasAssemblyPath" />
 		</MSBuild>
+		<MSBuild Projects="..\InfoPanel.StopWatch\InfoPanel.StopWatch.csproj" Targets="GetTargetPath" Properties="Configuration=$(Configuration);Platform=$(Platform);TargetFramework=net8.0-windows;RuntimeIdentifier=win-x64">
+			<Output TaskParameter="TargetOutputs" PropertyName="_StopWatchAssemblyPath" />
+		</MSBuild>
 		<PropertyGroup>
 			<ExtrasOutputDir>$([System.IO.Path]::GetDirectoryName('$(_ExtrasAssemblyPath)'))</ExtrasOutputDir>
+			<StopWatchOutputDir>$([System.IO.Path]::GetDirectoryName('$(_StopWatchAssemblyPath)'))</StopWatchOutputDir>
 		</PropertyGroup>
 		<MakeDir Directories="$(PublishDir)\plugins\InfoPanel.Extras" />
+		<MakeDir Directories="$(PublishDir)\plugins\InfoPanel.StopWatch" />
 		<ItemGroup>
 			<ExtrasFiles Include="$(ExtrasOutputDir)\**\*.*" />
+			<StopWatchFiles Include="$(StopWatchOutputDir)\**\*.*" />
 		</ItemGroup>
 		<Copy SourceFiles="@(ExtrasFiles)" DestinationFolder="$(PublishDir)\plugins\InfoPanel.Extras" SkipUnchangedFiles="true" />
+		<Copy SourceFiles="@(StopWatchFiles)" DestinationFolder="$(PublishDir)\plugins\InfoPanel.StopWatch" SkipUnchangedFiles="true" />
 	</Target>
 
 	<Target Name="CleanPluginsDirectory" BeforeTargets="Clean">
