+++ InfoPanel/Models/StopwatchHotkeySupport.cs
@@ -1,8 +1,11 @@
 namespace InfoPanel.Models
 {
-    /// <summary>Plugin id for the bundled Stopwatch plugin in InfoPanel.Extras (global hotkeys use this id).</summary>
+    /// <summary>Plugin id for the bundled InfoPanel.StopWatch plugin (global hotkeys use this id).</summary>
     internal static class StopwatchHotkeySupport
     {
         public const string PluginId = "infopanel-stopwatch";
+
+        /// <summary>Matches PluginInfo.ini Name for the bundled Stopwatch package (Plugins UI).</summary>
+        internal const string BundledPackageName = "InfoPanel.StopWatch";
     }
 }
